PMM-13746 resolve comments
