+++ public/app/percona/shared/components/PerconaBootstrapper/PerconaNewVersion/PerconaNewVersion.constants.ts
@@ -1,11 +1,2 @@
-export const Messages = {
-  titleNewVersion: 'PMM 3 is here and it’s better',
-  viewMigrationGuide: 'View migration guide',
-  close: 'Close',
-  description:
-    'Upgrade for enhanced security, improved stability, and a smoother experience. PMM 3 also sets the foundation for exciting future updates. ',
-  seeWhatsNew: 'See what’s new in PMM 3',
-};
-
 export const DOCUMENTATION_LINK_URL = 'https://per.co.na/pmm/3.0.0/release-notes';
 export const MIGRATION_LINK_URL = 'https://per.co.na/migrate_to_PMM3';
